Update index.md
When I was checking this page to try out toastify- https://fkhadra.github.io/react-toastify/introduction, I was trying to change the default Transition, but I was unable to do so. 
Then I had to dig in the deep to find the solution. 
So, I think adding the documentation link in this readme page will be helpful to the new users.